Design: 회원가입 뷰 UI 추가
